LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/metadata/generated/element-a.https.sub-expected.txt


FAIL sec-fetch-site - Same origin - no attributes assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
FAIL sec-fetch-site - Cross-site - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same site - no attributes assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect - no attributes assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Same Origin - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Same-Site - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Cross-Site -> Cross-Site - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same Origin - no attributes assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Same-Site - no attributes assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Origin -> Cross-Site - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Same Origin - no attributes assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Same-Site - no attributes assert_array_equals: expected property 0 to be "same-site" but got "none" (expected array ["same-site"] got ["none"])
FAIL sec-fetch-site - Same-Site -> Cross-Site - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
FAIL sec-fetch-site - HTTPS downgrade-upgrade - no attributes assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
PASS sec-fetch-mode - no attributes
FAIL sec-fetch-mode - attributes: download assert_own_property: expected property "sec-fetch-mode" missing
PASS sec-fetch-dest - no attributes
FAIL sec-fetch-dest - attributes: download assert_own_property: expected property "sec-fetch-dest" missing
PASS sec-fetch-user - no attributes
FAIL sec-fetch-user - no attributes with user activation assert_own_property: expected property "sec-fetch-user" missing
